Add skeleton for IR node for MultipleJoin
Lays out the skeleton for an IR node currently called
TableMultiOuterJoin that takes a list of tables and a field name.
All tables must have the same key and row type. The output type is
a Table with two fields. The key, and an array of structs that represent
the row data named `fieldName`.