Ensure that a block is always decompressed in virtualSeek
This fixes a small bug where after changing decompression to be lazy, if
virtualSeek is called with a pointer into the first block, it's possible
nothing has been decompressed yet. Add an extra check, so that the start
of the file is always decompressed. This may end up being a little extra
work if it's called over and over again, but in practice, it gets called
once per file, and then the work would need to be done anyways